Inform user before exiting preferences dialog with changes
This Gray out the apply button by default. If something change
the button apply is enabled.

If you try to exit the windows without saving you got a warning.

The change detection is made by binding to various signal of the
preferences dialog childrens.

This work also when coming from the setup wizard.

Fix #515