Show the server choice in wizard if you have a VM or a remote server
It's easier for using instead of searching the screen with use
local server options.

But I think it's not complete because we still have the checkbox
in the preferences.

When I read the code I found this:
https://github.com/GNS3/gns3-gui/blob/cd9bb16f79e8738fff34ff695730163b92c11170/gns3/modules/builtin/__init__.py#L72

And I'm not sure in which case it's call. I think it's only for
VPCS and things like Dynamips switch. Is it right? If it's the case
we can
drop the checkbox from IOU, Qemu, VMware, VirtualBox