Drop SSH support
Since we introduce the SSH support we got bug and bug
due to the complexity introduce by the tunnels. Even
in the 1.4.0 after long testing nobody notice that
the feature was broke.

And in his actual implementation we have limitation
preventing you to connect a node running on your PC
to a node running on the remote server.

After thinking about that it's better to drop it before
this feature is used by a lot of people. It's better to
use a VPN and less ninja.

Also we know that embeding paramiko require us to distribute
an encryption library creating legal trouble with US embargo
on various country.

Bonus point for the future: SSH support is probalby hard, or impossible for
a web interface and complicate the move to an architecture
with GNS3 server running as controller managing all the servers.

Fix #930