Support for profil settings
This PR add a --profil command line arguments.

If set a new set of settings will be create in
~/GNS3/profiles/PROFILENAME

Otherwise it's use the default location.

This allow to test settings without losing
original.

Also with the correct GUI this allow to switch from
local server to a remote (home / office).